feat(all): look for build.ghostdog instead of BUILD
BUILD files are already used by a bunch of other projects. ghostdog
isn't intended to compete with these other tools. In the future, I want
ghostdog to either co-exist or let ghostdog let folks gradually evolve
into using other tools. This will be easier if not competing on the
filename.