chore(_examples): add _ so go ignores this directory
Before, when running `make test-unit`, go would would run the unit tests
of examples/go/simple, which is not desired. Fortunately, prefixing a
directory with an underscore causes go to ignore this directory.

So, the build integration test had to be updated. Thought this was a bug
with ghostdog, but the `shasum` of the `main` created by `go build`
changes when the directory name was changed from examples to _examples.
Kind of a surprise to me. Guessing the go compiler caches based on
absolute filename.